+++ REST_API_SERVER/Database_Models/Employee.cs
@@ -18,7 +18,7 @@ namespace REST_API_SERVER.Database_Models
         public string SecondLastName { get; set; }
         public DateTime? FirstDateWorking { get; set; }
         public string CinemaName { get; set; }
-        public int? RoleId { get; set; }
+        public short? RoleId { get; set; }
 
         public virtual Cinema CinemaNameNavigation { get; set; }
         public virtual Role Role { get; set; }
